// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  MODERATOR: 'MODERATOR',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Visibility = {
  PRIVATE: 'PRIVATE',
  PUBLIC: 'PUBLIC'
} as const

export type Visibility = (typeof Visibility)[keyof typeof Visibility]


export const FlagStatus = {
  OPEN: 'OPEN',
  REVIEWED: 'REVIEWED',
  CLOSED: 'CLOSED'
} as const

export type FlagStatus = (typeof FlagStatus)[keyof typeof FlagStatus]